feat: get and confirm changes

+++ src/db/queries/net/circle.ts
@@ -123,7 +123,7 @@ export const getVotes = `
 export const getMembers = `
   SELECT
     nodes.*,
-    nets_users_data.user_id,
+    nets_users_data.user_id::int,
     nets_users_data.confirmed
   FROM nodes
   INNER JOIN nets_users_data ON
